BoxCoxTransformations.jl

Base.empty!

BoxCoxTransformations.transform

StatsAPI.confint

StatsAPI.fit

StatsAPI.loglikelihood

Base.empty! — Method

empty!(bc::PowerTransformation)

Remove data.

BoxCoxTransformations.transform — Method

transform(bc::BoxCoxTransformation; α = 0, scaled = false)
transform(bc::BoxCoxTransformation, λ::Real; α = 0, scaled = false)

Returns transformed data for optimal λ as specified in bc or for a particular λ, if specified.

Keyword arguments:

α: added to all values in 𝐱 before transformation. Default = 0.

scaled: scale transformation results. Default = false.

StatsAPI.confint — Method

confint(bc::BoxCoxTransformation; level=0.95)

Returns confidence interval for the estimated maximum power parameter λ of a given BoxCoxTransformation.

Note: The confidence interval is the range of λs for which the log-likelihood(λ) >= loglikelihood(λ_max) - 0.5 * quantile(Chisq(1), level)

Reference: Linear Models With R by Julian Faraway Section 7.1.

StatsAPI.fit — Method

fit(::BoxCoxTransformation, 𝐱::Vector{<:Real};
		interval = (-2.0, 2.0),method = :geomean, kwargs...)

Fit box cox transformation to data.

Power parameter, λ, is determined from 𝐱 using a log-likelihood estimator.

Keyword arguments:

method: for the cacluation of the log-likelihood

- :geomean => -N / 2.0 * log(2 * π * σ² / gm ^ (2 * (λ - 1)) + 1)
- :normal => -N / 2.0 * log(σ²) + (λ - 1) * sum(log.(𝐱))

any other keyword arguments accepted by Optim.optimize function e.g. abs_tol

StatsAPI.loglikelihood — Method

loglikelihood(bc::BoxCoxTransformation)
loglikelihood(bc::BoxCoxTransformation, λ::Real)

Returns log likelihood for optimal λ as specified in bc or for a particular λ, if specified.